+++ src/test/kotlin/com/terraformation/backend/tracking/db/T0StoreTest.kt
@@ -211,36 +211,6 @@ internal class T0StoreTest : DatabaseTest(), RunsAsDatabaseUser {
   }
 
   @Nested
-  inner class FetchSurvivalRateCalculationInProgress {
-    @Test
-    fun `throws exception when user lacks permission`() {
-      deleteOrganizationUser()
-
-      assertThrows<PlantingSiteNotFoundException> {
-        store.fetchSurvivalRateCalculationInProgress(plantingSiteId)
-      }
-    }
-
-    @Test
-    fun `returns false when no recalculation is in progress`() {
-      assertFalse(
-          store.fetchSurvivalRateCalculationInProgress(plantingSiteId),
-          "No recalculation marker row means no calculation in progress",
-      )
-    }
-
-    @Test
-    fun `returns true when a recalculation marker exists`() {
-      insertPlantingSiteSurvivalRateCalculation()
-
-      assertTrue(
-          store.fetchSurvivalRateCalculationInProgress(plantingSiteId),
-          "Recalculation marker row means a calculation is in progress",
-      )
-    }
-  }
-
-  @Nested
   inner class FetchAllT0SiteDataSet {
     @Test
     fun `throws exception when user lacks permission`() {
